Changed song ints to floats
In case a song is 120.5 bpm instead of 120.

+++ Assets/Scripts/SongController.cs
@@ -23,7 +23,7 @@ public class SongController : MonoBehaviour
     private float beatTimer;
     private float lateBeatTimer;
 
-    private int beatsUntilStart;
+    private float beatsUntilStart;
 
     private void Start()
     {
